# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: credit_card
  description: ''
  version: 1.0.0
...
paths:
...
  /payments/{paymentId}:
    delete:
      summary: closePayment
      deprecated: false
      description: ''
      operationId: closePayment
      tags: []
      parameters:
        - name: paymentId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /cards/{cardId}/payment:
    post:
      summary: createPayment
      deprecated: false
      description: ''
      operationId: createPayment
      tags: []
      parameters:
        - name: cardId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            examples: {}
        required: true
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
...
components:
  schemas:
...
    PaymentRequest:
      type: object
      properties:
        channel:
          type: string
          pattern: ^(WEB|APP)$
        currency:
          type: string
          pattern: ^(PEN|USD)$
        amount:
          type: number
        category:
          type: string
          pattern: ^(MES|ADELANTADO|TOTAL)$
      required:
        - amount
        - currency
        - channel
        - category
...
    Tracking:
      type: object
      properties:
        trackingId:
          type: string
          format: uuid
        operationDate:
          type: string
          format: date-time
      required:
        - trackingId
        - operationDate
    DefaultResponse4xx:
      type: object
      properties:
        description:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - description
        - errors
    DefaultResponse:
      type: object
      properties:
        description:
          type: string
      required:
        - description
  responses:
    BadRequest:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DefaultResponse4xx'
    NotFound:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DefaultResponse'
    InternalError:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DefaultResponse'
  securitySchemes: {}
servers: []
security: []